remove the prisoner's objective
