--- Content.Server/_Harmony/Objectives/Components/UnrestrainedConditionComponent.cs
@@ -1,9 +1,0 @@
-﻿using Content.Server._Harmony.Objectives.Systems;
-
-namespace Content.Server._Harmony.Objectives.Components;
-
-/// <summary>
-/// Requires that the player is unrestrained.
-/// </summary>
-[RegisterComponent, Access(typeof(UnrestrainedConditionSystem))]
-public sealed partial class UnrestrainedConditionComponent : Component;
