+++ Content.Server/_Harmony/Objectives/Systems/KeepUnrestrainedConditionSystem.cs
@@ -1,13 +1,18 @@
 ﻿using Content.Server._Harmony.Objectives.Components;
+using Content.Server.Cuffs;
+using Content.Server.Mind;
 using Content.Server.Objectives.Systems;
+using Content.Shared.Cuffs.Components;
+using Content.Shared.Mind;
 using Content.Shared.Objectives.Components;
 
 namespace Content.Server._Harmony.Objectives.Systems;
 
 public sealed class KeepUnrestrainedConditionSystem : EntitySystem
 {
+    [Dependency] private readonly CuffableSystem _cuffableSystem = default!;
+    [Dependency] private readonly MindSystem _mindSystem = default!;
     [Dependency] private readonly TargetObjectiveSystem _target = default!;
-    [Dependency] private readonly UnrestrainedConditionSystem _unrestrainedCondition = default!;
 
     public override void Initialize()
     {
@@ -23,6 +28,17 @@ public sealed class KeepUnrestrainedConditionSystem : EntitySystem
         if (!_target.GetTarget(uid, out var target))
             return;
 
-        args.Progress = _unrestrainedCondition.GetProgress((args.MindId, args.Mind));
+        args.Progress = GetProgress((args.MindId, args.Mind));
+    }
+
+    public float GetProgress(Entity<MindComponent> mind)
+    {
+        if (mind.Comp.OwnedEntity is not {} entity || _mindSystem.IsCharacterDeadIc(mind))
+            return 0f;
+
+        if (!TryComp<CuffableComponent>(entity, out var cuffable))
+            return 0f;
+
+        return _cuffableSystem.IsCuffed((entity, cuffable)) ? 0f : 1f;
     }
 }
