+++ Resources/Locale/en-US/_Harmony/perma-prisoner/role.ftl
@@ -1,3 +1,1 @@
 perma-prisoner-round-end-agent-name = prisoner
-
-objective-issuer-perma-prisoner = [color=#ff8300]Prisoner[/color]
